(Chore) Move user container components
