+++ src/signals/settings/users/Detail/__tests__/Detail.test.js
