+++ src/signals/settings/users/Detail/components/UserForm/__tests__/UserForm.test.js
